requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.34.127
botocore==1.34.127
Django==4.2.13
djangorestframework==3.15.1
dnspython==2.6.1
jmespath==1.0.1
pymongo==4.7.3
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
urllib3==1.26.18
